okay fixed var

+++ src/var.py
@@ -33,7 +33,7 @@ def lut_var(var, states):
 # https://stackoverflow.com/questions/20626994/how-to-calculate-the-inverse-of-the-normal-cumulative-distribution-function-in-p
 
 def lut_var(var, states):
-    lut = np.zeros(shape=(states + 1, 1000), dtype=np.int32)
+    lut = np.zeros(shape=(states + 1, 1000), dtype=np.float32)
     for s in range(1, states + 1):
         std = var * np.sqrt(s)
         # cannot do 0 or 1, these are np.inf.
@@ -41,7 +41,7 @@ def lut_var(var, states):
         minval = norm.cdf(-3, loc=0, scale=1)
         maxval = norm.cdf(3, loc=0, scale=1)
         lut[s, :] = norm.ppf(q=np.linspace(minval, maxval, 1000), loc=0., scale=std)
-        
+
     return lut
